+++ src/main/java/net/greatstart/dao/EventDao.java
@@ -1,0 +1,7 @@
+package net.greatstart.dao;
+
+import net.greatstart.model.Event;
+
+
+public interface EventDao extends GenericDAO<Event> {
+}
