+++ src/main/java/net/greatstart/dao/InvestorDao.java
@@ -1,0 +1,7 @@
+package net.greatstart.dao;
+
+import net.greatstart.model.Investor;
+
+
+public interface InvestorDao extends GenericDAO<Investor> {
+}
