+++ src/main/java/net/greatstart/dao/ProjectOwnerDao.java
@@ -1,0 +1,7 @@
+package net.greatstart.dao;
+
+import net.greatstart.model.ProjectOwner;
+
+
+public interface ProjectOwnerDao extends GenericDAO<ProjectOwner>{
+}
