+++ src/main/java/net/greatstart/dao/hibernate/EventDaoImpl.java
@@ -1,0 +1,35 @@
+package net.greatstart.dao.hibernate;
+
+import net.greatstart.dao.EventDao;
+import net.greatstart.model.Event;
+
+import java.util.List;
+
+
+public class EventDaoImpl implements EventDao {
+
+    public void create(Event item) {
+
+    }
+
+    public boolean delete(long id) {
+        return false;
+    }
+
+    public boolean update(Event item) {
+        return false;
+    }
+
+    public Event getByEmail(String email) {
+        return null;
+    }
+
+    public Event getById(long id) {
+        return null;
+    }
+
+    public List<Event> getAll() {
+        return null;
+    }
+
+}
