+++ src/main/java/net/greatstart/dao/hibernate/InvestorDaoImpl.java
@@ -1,0 +1,34 @@
+package net.greatstart.dao.hibernate;
+
+import net.greatstart.dao.InvestorDao;
+import net.greatstart.model.Investor;
+
+import java.util.List;
+
+
+public class InvestorDaoImpl implements InvestorDao {
+
+    public void create(Investor item) {
+
+    }
+
+    public boolean delete(long id) {
+        return false;
+    }
+
+    public boolean update(Investor item) {
+        return false;
+    }
+
+    public Investor getByEmail(String email) {
+        return null;
+    }
+
+    public Investor getById(long id) {
+        return null;
+    }
+
+    public List<Investor> getAll() {
+        return null;
+    }
+}
