+++ src/main/java/net/greatstart/dao/hibernate/ProjectOwnerDaoImpl.java
@@ -1,0 +1,35 @@
+package net.greatstart.dao.hibernate;
+
+import net.greatstart.dao.ProjectOwnerDao;
+import net.greatstart.model.ProjectOwner;
+
+import java.util.List;
+
+
+public class ProjectOwnerDaoImpl implements ProjectOwnerDao {
+
+    public void create(ProjectOwner item) {
+
+    }
+
+    public boolean delete(long id) {
+        return false;
+    }
+
+    public boolean update(ProjectOwner item) {
+        return false;
+    }
+
+    public ProjectOwner getByEmail(String email) {
+        return null;
+    }
+
+    public ProjectOwner getById(long id) {
+        return null;
+    }
+
+    public List<ProjectOwner> getAll() {
+        return null;
+    }
+
+}
