+++ src/main/java/net/greatstart/filters/RequestFilter.java
@@ -1,0 +1,20 @@
+package net.greatstart.filters;
+
+import javax.servlet.*;
+import java.io.IOException;
+
+
+public class RequestFilter implements Filter {
+
+    public void init(FilterConfig filterConfig) throws ServletException {
+
+    }
+
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
+
+    }
+
+    public void destroy() {
+
+    }
+}
