+++ src/main/java/net/greatstart/model/Cards/Project.java
@@ -1,0 +1,18 @@
+package net.greatstart.model.Cards;
+
+import net.greatstart.model.ProjectOwner;
+
+import java.util.Date;
+
+public class Project {
+
+    private boolean isVerifed;
+    private String name;
+    private ProjectDescription desc;
+    private ProjectOwner owner;
+    private int cost;
+    private int minInvestment;
+    private Date addDate;
+//    later: int rating (view counter)
+//    URl logo
+}
