+++ src/main/java/net/greatstart/model/Cards/ProjectDescription.java
@@ -1,0 +1,12 @@
+package net.greatstart.model.Cards;
+
+import java.util.Date;
+
+public class ProjectDescription {
+
+    //    Url url
+    private Date date;
+    private String description;
+    private String goal;
+    private String other;
+}
