+++ src/main/java/net/greatstart/model/Contact.java
@@ -1,0 +1,12 @@
+package net.greatstart.model;
+
+
+public class Contact {
+    private String address;
+    private long phoneNumber;
+
+    public Contact(String address, long phoneNumber) {
+        this.address = address;
+        this.phoneNumber = phoneNumber;
+    }
+}
