+++ src/main/java/net/greatstart/model/Event.java
@@ -1,0 +1,11 @@
+package net.greatstart.model;
+
+import java.util.Date;
+
+public class Event {
+    private String name;
+
+    private String description;
+
+    private Date date;
+}
