+++ src/main/java/net/greatstart/model/Investment.java
@@ -1,0 +1,14 @@
+package net.greatstart.model;
+
+import net.greatstart.model.Cards.Project;
+
+import java.math.BigDecimal;
+import java.util.Date;
+
+public class Investment {
+
+    private Date dateOfInvestment;
+    private Project project;
+    private Investor inv;
+    private BigDecimal sum;
+}
