+++ src/main/java/net/greatstart/model/Role.java
@@ -1,0 +1,15 @@
+package net.greatstart.model;
+
+public enum Role {
+    USER("user"), ADMIN("admin");
+
+    private String name;
+
+    private Role(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
