--- src/main/java/ua/goit/group6/dao/EventDao.java
@@ -1,7 +1,0 @@
-package ua.goit.group6.dao;
-
-import ua.goit.group6.model.Event;
-
-
-public interface EventDao extends GenericDAO<Event> {
-}
