--- src/main/java/ua/goit/group6/dao/InvestorDao.java
@@ -1,7 +1,0 @@
-package ua.goit.group6.dao;
-
-import ua.goit.group6.model.Investor;
-
-
-public interface InvestorDao extends GenericDAO<Investor> {
-}
