--- src/main/java/ua/goit/group6/dao/ProjectOwnerDao.java
@@ -1,7 +1,0 @@
-package ua.goit.group6.dao;
-
-import ua.goit.group6.model.ProjectOwner;
-
-
-public interface ProjectOwnerDao extends GenericDAO<ProjectOwner>{
-}
