--- src/main/java/ua/goit/group6/dao/hibernate/EventDaoImpl.java
@@ -1,35 +1,0 @@
-package ua.goit.group6.dao.hibernate;
-
-import ua.goit.group6.dao.EventDao;
-import ua.goit.group6.model.Event;
-
-import java.util.List;
-
-
-public class EventDaoImpl implements EventDao {
-
-    public void create(Event item) {
-
-    }
-
-    public boolean delete(long id) {
-        return false;
-    }
-
-    public boolean update(Event item) {
-        return false;
-    }
-
-    public Event getByEmail(String email) {
-        return null;
-    }
-
-    public Event getById(long id) {
-        return null;
-    }
-
-    public List<Event> getAll() {
-        return null;
-    }
-
-}
