--- src/main/java/ua/goit/group6/dao/hibernate/InvestorDaoImpl.java
@@ -1,34 +1,0 @@
-package ua.goit.group6.dao.hibernate;
-
-import ua.goit.group6.dao.InvestorDao;
-import ua.goit.group6.model.Investor;
-
-import java.util.List;
-
-
-public class InvestorDaoImpl implements InvestorDao {
-
-    public void create(Investor item) {
-
-    }
-
-    public boolean delete(long id) {
-        return false;
-    }
-
-    public boolean update(Investor item) {
-        return false;
-    }
-
-    public Investor getByEmail(String email) {
-        return null;
-    }
-
-    public Investor getById(long id) {
-        return null;
-    }
-
-    public List<Investor> getAll() {
-        return null;
-    }
-}
