--- src/main/java/ua/goit/group6/dao/hibernate/ProjectOwnerDaoImpl.java
@@ -1,35 +1,0 @@
-package ua.goit.group6.dao.hibernate;
-
-import ua.goit.group6.dao.ProjectOwnerDao;
-import ua.goit.group6.model.ProjectOwner;
-
-import java.util.List;
-
-
-public class ProjectOwnerDaoImpl implements ProjectOwnerDao {
-
-    public void create(ProjectOwner item) {
-
-    }
-
-    public boolean delete(long id) {
-        return false;
-    }
-
-    public boolean update(ProjectOwner item) {
-        return false;
-    }
-
-    public ProjectOwner getByEmail(String email) {
-        return null;
-    }
-
-    public ProjectOwner getById(long id) {
-        return null;
-    }
-
-    public List<ProjectOwner> getAll() {
-        return null;
-    }
-
-}
