--- src/main/java/ua/goit/group6/filters/FilterAuthorization.java
@@ -1,19 +1,0 @@
-package ua.goit.group6.filters;
-
-import javax.servlet.*;
-import java.io.IOException;
-
-public class FilterAuthorization implements Filter{
-
-    public void init(FilterConfig filterConfig) throws ServletException {
-
-    }
-
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-
-    }
-
-    public void destroy() {
-
-    }
-}
