--- src/main/java/ua/goit/group6/filters/RequestFilter.java
@@ -1,20 +1,0 @@
-package ua.goit.group6.filters;
-
-import javax.servlet.*;
-import java.io.IOException;
-
-
-public class RequestFilter implements Filter {
-
-    public void init(FilterConfig filterConfig) throws ServletException {
-
-    }
-
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-
-    }
-
-    public void destroy() {
-
-    }
-}
