--- src/main/java/ua/goit/group6/model/Cards/Project.java
@@ -1,18 +1,0 @@
-package ua.goit.group6.model.Cards;
-
-import ua.goit.group6.model.ProjectOwner;
-
-import java.util.Date;
-
-public class Project {
-
-    private boolean isVerifed;
-    private String name;
-    private ProjectDescription desc;
-    private ProjectOwner owner;
-    private int cost;
-    private int minInvestment;
-    private Date addDate;
-//    later: int rating (view counter)
-//    URl logo
-}
