--- src/main/java/ua/goit/group6/model/Cards/ProjectDescription.java
@@ -1,12 +1,0 @@
-package ua.goit.group6.model.Cards;
-
-import java.util.Date;
-
-public class ProjectDescription {
-
-    //    Url url
-    private Date date;
-    private String description;
-    private String goal;
-    private String other;
-}
