--- src/main/java/ua/goit/group6/model/Contact.java
@@ -1,12 +1,0 @@
-package ua.goit.group6.model;
-
-
-public class Contact {
-    private String address;
-    private long phoneNumber;
-
-    public Contact(String address, long phoneNumber) {
-        this.address = address;
-        this.phoneNumber = phoneNumber;
-    }
-}
