--- src/main/java/ua/goit/group6/model/Event.java
@@ -1,11 +1,0 @@
-package ua.goit.group6.model;
-
-import java.util.Date;
-
-public class Event {
-    private String name;
-
-    private String description;
-
-    private Date date;
-}
