--- src/main/java/ua/goit/group6/model/Investment.java
@@ -1,14 +1,0 @@
-package ua.goit.group6.model;
-
-import ua.goit.group6.model.Cards.Project;
-
-import java.math.BigDecimal;
-import java.util.Date;
-
-public class Investment {
-
-    private Date dateOfInvestment;
-    private Project project;
-    private Investor inv;
-    private BigDecimal sum;
-}
