+++ src/test/java/net/greatstart/model/Test.java
@@ -1,0 +1,7 @@
+package net.greatstart.model;
+
+/**
+ * Created by Garik on 10.04.2017.
+ */
+public class Test {
+}
