--- src/test/java/ua/goit/group6/model/Test.java
@@ -1,7 +1,0 @@
-package ua.goit.group6.model;
-
-/**
- * Created by Garik on 10.04.2017.
- */
-public class Test {
-}
